Rename projects cache helper module
