+++ hippo-frontend/src/graphql/helpers/project_cache.js
